[Update] 재연결 기능 업데이트
